add few code for provider

+++ internal/domain/entity.go
@@ -4,7 +4,7 @@ type MarketTicker struct {
 	// 8 bytes group
 	Price     float64
 	Volume    float64
-	Timestamp float64
+	Timestamp int64
 
 	// 16 bytes group
 	Symbole  string
